+++ back/src/main/java/com/baba/back/content/dto/CreateCommentRequest.java
@@ -1,0 +1,18 @@
+package com.baba.back.content.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateCommentRequest {
+
+    @NotNull
+    private String tag;
+
+    @NotNull
+    private String comment;
+}
