+++ back/src/main/java/com/baba/back/content/service/ContentService.java
@@ -3,7 +3,7 @@ package com.baba.back.content.service;
 import com.baba.back.baby.domain.Baby;
 import com.baba.back.baby.exception.BabyNotFoundException;
 import com.baba.back.baby.repository.BabyRepository;
-import com.baba.back.content.domain.FileHandler;
+import com.baba.back.common.FileHandler;
 import com.baba.back.content.domain.ImageFile;
 import com.baba.back.content.domain.Like;
 import com.baba.back.content.domain.content.Content;
